refactor(soffit-pwa): ui folder structure
