+++ uniquely/soffit-pwa/src/ui/stuff/app-footer.ts
