+++ uniquely/soffit-pwa/src/ui/stuff/supply-chain-box.ts
